requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
aioredis==2.0.1
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1